+++ .script/lib/sh/project.sh
@@ -2,12 +2,11 @@
 # Copyright (C) 2019 Ethan Pini <epini@sfu.ca>
 # MIT License
 # ----------------------------------------------------------------------------------------------------------------------
-# A function for lazily loading the git utilities.
+# A function for querying project information.
+# This defers all processing to libexec/project.js
 # ----------------------------------------------------------------------------------------------------------------------
 
-git_utils() {
-	local util
-	for util in "${BASE}/lib/git_"*.sh; do
-		source "$util"
-	done
+project() {
+	"${BASE}/libexec/project.js" "$@"
+	return $?
 }